+++ src/Tests/Unit/Private/Invoke-XMLDataCheck.Tests.ps1
@@ -35,34 +35,34 @@ InModuleScope 'PSGalleryExplorer' {
             } #before_each
             It 'Should process by default' {
                 Invoke-XMLDataCheck
-                Assert-MockCalled Invoke-XMLDataCheck -Scope It -Exactly -Times 1
+                Should -Invoke Invoke-XMLDataCheck -Scope It -Exactly -Times 1
             } #it
             It 'Should not process on explicit request for confirmation (-Confirm)' {
                 { Invoke-XMLDataCheck -Confirm }
-                Assert-MockCalled Invoke-XMLDataCheck -Scope It -Exactly -Times 0
+                Should -Invoke Invoke-XMLDataCheck -Scope It -Exactly -Times 0
             } #it
             It 'Should not process on implicit request for confirmation (ConfirmPreference)' {
                 {
                     $ConfirmPreference = 'Low'
                     Invoke-XMLDataCheck
                 }
-                Assert-MockCalled Invoke-XMLDataCheck -Scope It -Exactly -Times 0
+                Should -Invoke Invoke-XMLDataCheck -Scope It -Exactly -Times 0
             } #it
             It 'Should not process on explicit request for validation (-WhatIf)' {
                 { Invoke-XMLDataCheck -WhatIf }
-                Assert-MockCalled Invoke-XMLDataCheck -Scope It -Exactly -Times 0
+                Should -Invoke Invoke-XMLDataCheck -Scope It -Exactly -Times 0
             } #it
             It 'Should not process on implicit request for validation (WhatIfPreference)' {
                 {
                     $WhatIfPreference = $true
                     Invoke-XMLDataCheck
                 }
-                Assert-MockCalled Invoke-XMLDataCheck -Scope It -Exactly -Times 0
+                Should -Invoke Invoke-XMLDataCheck -Scope It -Exactly -Times 0
             } #it
             It 'Should process on force' {
                 $ConfirmPreference = 'Medium'
                 Invoke-XMLDataCheck -Force
-                Assert-MockCalled Invoke-XMLDataCheck -Scope It -Exactly -Times 1
+                Should -Invoke Invoke-XMLDataCheck -Scope It -Exactly -Times 1
             } #it
         } #context_shouldprocess