+++ docs/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - SSM Task now generates metadata file
     - InvokeBuild bumped from `5.10.2` to `5.10.3`
     - Convert bumped from `1.2.0` to `1.5.0`
+    - Removed all test case uses of `Assert-MockCalled`
 
 ## [2.1.0]